+++ src/test/resources/stage.properties
@@ -12,7 +12,7 @@ us16Subject=Lost Bag
 us16Message=My bag was lost on expedition 234532. A red medium-sized bag.
 
 usernameRegistered=logintesti
-passwordRegistered=LoginTesti!9
+passwordRegistered=LoginTesti!10
 changedPassword=LoginTesti!10
 confirmPassword=LoginTesti!10
 imgPathAttachment = src/test/resources/testResources/attachment.png
